@@ -1,4 +1,6 @@
-public struct InvocationResponse<T>: Codable where T: Codable
+import Foundation
+
+public struct InvocationResponse<T>: Codable, Sendable where T: Codable & Sendable
 {
     var error: String?
     var value: T?
@@ -15,4 +17,26 @@ extension InvocationResponse where T == Never
 extension InvocationResponse where T == Never
 {
     public init() {}
+}
+
+extension InvocationResponse where T == Data
+{
+    public init(error: Swift.Error)
+    {
+        self.error = String(describing: error)
+    }
+}
+
+extension InvocationResponse where T == Data
+{
+    public init(value: Data)
+    {
+        self.value = value
+    }
+    
+    public init(error: String?, value: Data?)
+    {
+        self.error = error
+        self.value = value
+    }
 }
